Remove AAC ViewModel!
