+++ build.gradle
@@ -5,13 +5,11 @@ buildscript {
         retrofit_version = "2.9.0"
         glide_version = "4.13.2"
         room_version = "2.4.3"
-        lifecycle_version = "2.5.1"
         coroutines_version = "1.6.4"
         mockk_version = "1.13.2"
         android_core_testing_version = "2.1.0"
         dagger_version = "2.44"
-        compose_version = "1.3.0"
-        hilt_version = "1.0.0"
+        compose_version = "1.3.1"
         work_version = "2.7.1"
     }
 }
